+++ EntityArchitect.CRUD/ApiBuilder.cs
@@ -1,6 +1,7 @@
 using System.Reflection;
 using System.Text.RegularExpressions;
 using EntityArchitect.CRUD.Attributes;
+using EntityArchitect.CRUD.Attributes.CrudAttributes;
 using EntityArchitect.CRUD.Authorization;
 using EntityArchitect.CRUD.Authorization.Attributes;
 using EntityArchitect.CRUD.Authorization.Service;
@@ -24,12 +25,25 @@ public static partial class ApiBuilder
     {
         var enumerable = assembly.ExportedTypes.Where(c => c.BaseType == typeof(Entity)).ToList();
         var typeBuilder = new TypeBuilder();
-        
         foreach (var entity in enumerable)
         {
             var result = ConvertEndpointNameRegex().Replace(entity.Name, "$1-$2");
             var name = result.ToLower();
 
+            var authorizationPolicies = new List<Type>();
+            var haveAuthorization = entity.CustomAttributes.Any(c => c.AttributeType == typeof(AuthorizationEntityAttribute));
+            var authorizationEntityAttribute = entity.GetCustomAttribute<AuthorizationEntityAttribute>();
+            if (authorizationEntityAttribute is not null)
+            {
+                foreach (var type in authorizationEntityAttribute.EntityTypes)
+                {
+                    if(type.BaseType != typeof(AuthorizationEntityAttribute))
+                        throw new Exception($"AuthorizationEntityAttribute can only have AuthorizationEntityAttribute as EntityTypes. {type.Name}");
+                        
+                    authorizationPolicies.Add(type);
+                }
+            }
+            
             var requestPostType = typeBuilder.BuildCreateRequestFromEntity(entity);
             var requestUpdateType = typeBuilder.BuildUpdateRequestFromEntity(entity);
             var responseType = typeBuilder.BuildResponseFromEntity(entity);
@@ -58,7 +72,10 @@ public static partial class ApiBuilder
                     var postHandler =
                         delegateBuilder!.GetType().GetProperty("PostDelegate")!.GetValue(delegateBuilder) as Delegate;
                     var endpoint = group.MapPost("", postHandler!);
-
+                    
+                    if(haveAuthorization)
+                        endpoint.RequireAuthorization(authorizationPolicies.Select(c => c.Name).ToArray());
+                    
                     endpoint.WithSummary($"Create {entity.Name}");
                     endpoint.WithDisplayName($"Create {entity.Name}");
                     endpoint.Produces(200, typeof(Result<>).MakeGenericType(responseType));
@@ -72,6 +89,9 @@ public static partial class ApiBuilder
                         delegateBuilder!.GetType().GetProperty("UpdateDelegate")!.GetValue(delegateBuilder) as Delegate;
                     var endpoint = group.MapPut("", updateHandler!);
 
+                    if(haveAuthorization)
+                        endpoint.RequireAuthorization(authorizationPolicies.Select(c => c.Name).ToArray());
+                    
                     endpoint.WithSummary($"Update {entity.Name}");
                     endpoint.WithDisplayName($"Update {entity.Name}");
                     endpoint.Produces(200, typeof(Result<>).MakeGenericType(responseType));
@@ -85,6 +105,9 @@ public static partial class ApiBuilder
                         delegateBuilder!.GetType().GetProperty("DeleteDelegate")!.GetValue(delegateBuilder) as Delegate;
                     var endpoint = group.MapDelete("{id}", deleteHandler!);
 
+                    if(haveAuthorization)
+                        endpoint.RequireAuthorization(authorizationPolicies.Select(c => c.Name).ToArray());
+                    
                     endpoint.WithSummary($"Delete {entity.Name} by Id");
                     endpoint.WithDisplayName($"Delete {entity.Name} by Id");
                     endpoint.Produces(200, typeof(Result));
@@ -105,6 +128,9 @@ public static partial class ApiBuilder
 
                     var endpoint = group.MapGet("{id}", getByIdHandler!);
 
+                    if(haveAuthorization)
+                        endpoint.RequireAuthorization(authorizationPolicies.Select(c => c.Name).ToArray());
+                    
                     endpoint.WithSummary($"Get {entity.Name} by Id");
                     endpoint.WithDisplayName($"Get {entity.Name} by Id");
                     endpoint.Produces(200, typeof(Result<>).MakeGenericType(responseType));
@@ -126,13 +152,17 @@ public static partial class ApiBuilder
                                 .Contains(typeof(LightListPropertyAttribute)))
                             .Select(c => c.Name)
                             .ToList();
-
+                    
                     var getLightListDelegate =
                         delegateBuilder!.GetType().GetProperty("GetLightListDelegate")!.GetValue(delegateBuilder) as
                             Delegate;
+                    
                     var endpoint = group.MapGet("light-list", getLightListDelegate!);
                     endpoint.WithSummary(
                         $"Get light list of {entity.Name}s. Only includes Id and {string.Join(",", lightListProperties)}");
+                    
+                    if(haveAuthorization)
+                        endpoint.RequireAuthorization(authorizationPolicies.Select(c => c.Name).ToArray());
                 }
 
                 if (entity.CustomAttributes.Any(c => c.AttributeType == typeof(GetListPaginatedAttribute)))
@@ -141,6 +171,10 @@ public static partial class ApiBuilder
                         delegateBuilder!.GetType().GetProperty("GetListDelegate")!
                             .GetValue(delegateBuilder) as Delegate;
                     var endpoint = group.MapGet("list/{page}", getLightListDelegate!);
+                    
+                    if(haveAuthorization)
+                        endpoint.RequireAuthorization(authorizationPolicies.Select(c => c.Name).ToArray());
+                    
                     endpoint.WithSummary($"Get list of {entity.Name}s paginated");
                     endpoint.WithOpenApi(op =>
                     {
@@ -197,7 +231,6 @@ public static partial class ApiBuilder
                         .GetMethod("MapGetEndpoint")?
                         .MakeGenericMethod(queryType, query, entity);
                     
-
                     mi!.Invoke(group, new object[] { group, query.Name, app});
                 }
 
@@ -223,6 +256,23 @@ public static partial class ApiBuilder
             var result = queryHandler.HandleAsync(query, param, connectionString, cancellationToken: default);
             return result;
         });
+        
+        var authorizationPolicies = new List<Type>();
+        var haveAuthorization = typeof(TQuery).CustomAttributes.Any(c => c.AttributeType == typeof(AuthorizationEntityAttribute));
+        var authorizationEntityAttribute = typeof(TQuery).GetCustomAttribute<AuthorizationEntityAttribute>();
+        if (authorizationEntityAttribute is not null)
+        {
+            foreach (var type in authorizationEntityAttribute.EntityTypes)
+            {
+                if(type.BaseType != typeof(AuthorizationEntityAttribute))
+                    throw new Exception($"AuthorizationEntityAttribute can only have AuthorizationEntityAttribute as EntityTypes. {type.Name}");
+                        
+                authorizationPolicies.Add(type);
+            }
+        }
+        
+        if(haveAuthorization)
+            endpoint.RequireAuthorization(authorizationPolicies.Select(c => c.Name).ToArray());
     }
 
     [GeneratedRegex("([a-z])([A-Z])")]
