+++ EntityArchitect.CRUD/Attributes/CrudAttributes/CannotCreateAttribute.cs
@@ -1,0 +1,3 @@
+namespace EntityArchitect.CRUD.Attributes.CrudAttributes;
+
+public class CannotCreateAttribute : Attribute;
