+++ EntityArchitect.CRUD/Attributes/CrudAttributes/CannotDeleteAttribute.cs
@@ -1,0 +1,3 @@
+namespace EntityArchitect.CRUD.Attributes.CrudAttributes;
+
+public class CannotDeleteAttribute : Attribute;
