+++ EntityArchitect.CRUD/Attributes/CrudAttributes/CannotGetByIdAttribute.cs
@@ -1,0 +1,3 @@
+namespace EntityArchitect.CRUD.Attributes.CrudAttributes;
+
+public class CannotGetByIdAttribute : Attribute;
