+++ EntityArchitect.CRUD/Attributes/CrudAttributes/CannotUpdateAttribute.cs
@@ -1,0 +1,3 @@
+namespace EntityArchitect.CRUD.Attributes.CrudAttributes;
+
+public class CannotUpdateAttribute : Attribute;
