+++ EntityArchitect.CRUD/Attributes/CrudAttributes/GetListPaginatedAttribute.cs
@@ -1,0 +1,3 @@
+namespace EntityArchitect.CRUD.Attributes.CrudAttributes;
+
+public class GetListPaginatedAttribute(int itemCount = 10) : Attribute;
