+++ EntityArchitect.CRUD/Attributes/CrudAttributes/HasLightListAttribute.cs
@@ -1,0 +1,3 @@
+namespace EntityArchitect.CRUD.Attributes.CrudAttributes;
+
+public class HasLightListAttribute : Attribute;
