+++ EntityArchitect.CRUD/Attributes/CrudAttributes/IncludeInGetAttribute.cs
@@ -1,0 +1,3 @@
+namespace EntityArchitect.CRUD.Attributes.CrudAttributes;
+
+public class IncludeInGetAttribute(int includingDeep = 0) : Attribute;
