+++ EntityArchitect.CRUD/Attributes/CrudAttributes/LightListPropertyAttribute.cs
@@ -1,0 +1,3 @@
+namespace EntityArchitect.CRUD.Attributes.CrudAttributes;
+
+public class LightListPropertyAttribute : Attribute;
