+++ EntityArchitect.CRUD/Authorization/Attributes/AuthorizationEntityAttribute.cs
@@ -3,18 +3,10 @@ namespace EntityArchitect.CRUD.Authorization.Attributes;
 [AttributeUsage(AttributeTargets.Class)]
 public class AuthorizationEntityAttribute : Attribute
 {
-    public string? Name { get; }
+    public Type[] EntityTypes { get; }
 
-    public AuthorizationEntityAttribute(string? name = null)
+    public AuthorizationEntityAttribute(params Type[] entityTypes)
     {
-        Name = name ?? GetDefaultName();
-    }
-
-    private static string GetDefaultName()
-    {
-        var stackTrace = new System.Diagnostics.StackTrace();
-        var frame = stackTrace.GetFrame(1);
-        var method = frame?.GetMethod();
-        return method?.DeclaringType?.Name ?? "UnknownClass";
+        EntityTypes = entityTypes;
     }
 }
