+++ EntityArchitect.CRUD/TypeBuilders/TypeBuilder.cs
@@ -2,6 +2,7 @@ using System.Reflection;
 using System.Reflection.Emit;
 using System.Text.RegularExpressions;
 using EntityArchitect.CRUD.Attributes;
+using EntityArchitect.CRUD.Attributes.CrudAttributes;
 using EntityArchitect.CRUD.Attributes.QueryResponseTypeAttributes;
 using EntityArchitect.CRUD.Queries;
 using EntityArchitect.Entities.Attributes;
