+++ EntityArchitect.Example/Entities/Client.cs
@@ -1,4 +1,5 @@
 using EntityArchitect.CRUD.Attributes;
+using EntityArchitect.CRUD.Attributes.CrudAttributes;
 using EntityArchitect.CRUD.Authorization;
 using EntityArchitect.CRUD.Authorization.Attributes;
 using EntityArchitect.Entities.Attributes;
